python/requirements.txt
colorama
PyYAML
opencv-python
lxml
tqdm
tensorflow==2.3.0rc0
easydict
matplotlib
pillow
torch==1.8.1+cu111
torchvision==0.9.1+cu111
